Redirect to home after admin login
Because the admin login is shown only in the Steam login page, it would
default to showing the Steam login page after succesful admin login,
which doesn't make too much sense. Also changed the auth interceptor to
redirect to the Steam login page by default.

Transition to the Steam login state by default, if not authenticated and
insufficient privileges are found.